+++ src/main/java/net/frozenblock/wilderwild/registry/WWSoundTypes.java
@@ -220,13 +220,6 @@ public final class WWSoundTypes {
 		WWSounds.BLOCK_MAPLE_LEAVES_HIT,
 		WWSounds.BLOCK_MAPLE_LEAVES_FALL
 	);
-	public static final SoundType FLOWER = new SoundType(0.8F, 1F,
-		WWSounds.BLOCK_FLOWER_BREAK,
-		WWSounds.BLOCK_FLOWER_STEP,
-		WWSounds.BLOCK_FLOWER_PLACE,
-		WWSounds.BLOCK_FLOWER_HIT,
-		WWSounds.BLOCK_FLOWER_FALL
-	);
 	public static final SoundType LILY_PAD = new SoundType(1F, 1F,
 		SoundEvents.BIG_DRIPLEAF_BREAK,
 		SoundEvents.BIG_DRIPLEAF_STEP,