+++ CHANGELOG.md
@@ -3,3 +3,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
 - Added unique sounds for Maple Leaves.
 - Fixed Pink Petals incorrectly being assigned a different block SoundType.
+- Removed duplicate flower sounds.